add master password authentification

+++ srcs/cryptage.cpp
@@ -7,7 +7,7 @@ std::string Cryptage::crypter(std::string texte, std::string cle)
 
     encryptedMessage = texte;
     
-    for (int i = 0; i < encryptedMessage.size(); i++) {
+    for (size_t i = 0; i < encryptedMessage.size(); i++) {
         encryptedMessage[i] = encryptedMessage[i] + (cle[i % cle.size()]);
     }
 
@@ -21,7 +21,7 @@ std::string Cryptage::decrypter(std::string texte, std::string cle)
 
     decryptedMessage = texte;
 
-    for (int i = 0; i < decryptedMessage.size(); i++) {
+    for (size_t i = 0; i < decryptedMessage.size(); i++) {
         decryptedMessage[i] = decryptedMessage[i] - (cle[i % cle.size()]);
     }
